Add files via upload
This includes a working version which will go out on the internet, or on the local PC directory, and retrieve an m3u8 file.  The class hierarchy for Playlist, VariantPlaylist, and MasterPlaylist has been added.  Once the m3u8 file is retrieved it is loaded into a "content list", and the createPlaylist function creates either a master or variant playlist object.  Currently, nothing is done with the playlist, but a print function has been added which verifies the playlist.contents list correct.