Add files via upload
The errors in this version are corrected, and playlist.checkResults[0] contains a validated tag for whether #EXTM3U is at the first line of the file.  The ValidationReport classes are not being used at this time, and the results are being stored inside the playlist objects, as they are validation objects anyways.