Add files via upload
Cleaned up the main function to just have one block where the checks are run, as the Visitor needs to decide which version of a check to implement not the main program.  Only the command line version is developed right now, with just the first line check.  A simple report print out is added for debugging purposes, but this will be changed to a function later on.