Add files via upload
There was a problem where the playlists were not being created properly in the event it was Master.  The issue was that the variant playlists were not being created properly (all their contents were appended etc.)  The issue was found to be that in createVariant() the content list [] was not initialized etc.  Multiple logging.info() statements have been added, and the content lists have been renamed to mContent[] and vContent[] for Master and Variant playlists.