Add files via upload
The VerCompatCheck validator checks for version number, and performs various tag checks for both Variant and Master playlists.  This compatibility check has been updated to now include the case where a master playlist is given, and a list of variants must now be verified as well.  Verified to work on both a web URL and local file.  Logging statements added for clarity in troubleshooting.