Add files via upload
Adding the clearMaster() function definition.  This function will clear the lists associated with a PlayList and MasterPlayList, and will be called by createMaster() eventually.  This will resolve the issue where successive reports are run on the same file, and the check contents grow instead of only reflecting the current run.