Add files via upload
Added some formatting to the output file for batch mode operation to make it easier to read.  Also, fixed a bug where if the batchFile.txt has multiple files on separate lines the readline() method for the file picks up the carriage return as well.  Used a strip('\n') function to resolve this on the inputLIne.